+++ gui/utils/stream_redirect.py
@@ -1,0 +1,13 @@
+from PyQt5.QtCore import QObject, pyqtSignal
+
+
+class Stream(QObject):
+    """Redirect script output to GUI"""
+
+    newText = pyqtSignal(str)
+
+    def write(self, text):
+        self.newText.emit(str(text))
+
+    def flush(self):
+        pass
